Authentication files refactor
