+++ src/main/java/com/example/wdmsystem/auth/AuthenticationController.java
@@ -1,6 +1,5 @@
-package com.example.wdmsystem.employee.system.authentication;
+package com.example.wdmsystem.auth;
 
-import com.example.wdmsystem.util.JwtUtil;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
