+++ src/main/java/com/example/wdmsystem/auth/Login.java
@@ -1,4 +1,4 @@
-package com.example.wdmsystem.employee.system.authentication;
+package com.example.wdmsystem.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
