+++ src/main/java/com/example/wdmsystem/auth/LoginResponse.java
@@ -1,0 +1,4 @@
+package com.example.wdmsystem.auth;
+
+public record LoginResponse(String token) {
+}
