--- src/main/java/com/example/wdmsystem/employee/system/authentication/LoginResponse.java
@@ -1,4 +1,0 @@
-package com.example.wdmsystem.employee.system.authentication;
-
-public record LoginResponse(String token) {
-}
